+++ libs/button/src/lib/button-core/button-core.ts
@@ -20,17 +20,23 @@ import {
 export const buttonClassName = 'spy-button-core';
 
 @Injectable()
-export abstract class ButtonCoreInputs implements AfterViewInit, OnChanges {
+export class ButtonCoreInputs {
   @Input() shape: ButtonShape = ButtonShape.Default;
   @Input() size: ButtonSize = ButtonSize.Medium;
   @Input() variant: ButtonVariant = ButtonVariant.Primary;
   @Input() @ToJson() attrs?: ButtonAttributes;
+}
 
+@Injectable()
+export abstract class ButtonCore extends ButtonCoreInputs
+  implements AfterViewInit, OnChanges {
   @ViewChild('buttonRef') buttonRef?: ElementRef;
 
   protected abstract buttonClassName: string;
 
-  constructor(private renderer: Renderer2) {}
+  constructor(public renderer: Renderer2) {
+    super();
+  }
 
   ngAfterViewInit(): void {
     this.setClassList();