+++ libs/button/src/lib/button-link/button-link.component.stories.ts
@@ -22,6 +22,10 @@ export const primary = (): IStory => ({
       { Primary: 'primary', Secondary: 'secondary', Critical: 'critical' },
       'primary',
     ),
-    size: select('Size', { Large: 'lg', Medium: 'md', Small: 'sm' }, 'lg'),
+    size: select(
+      'Size',
+      { Large: 'lg', Medium: 'md', Small: 'sm', ExtraSmall: 'xs' },
+      'lg',
+    ),
   },
 });