+++ libs/button/src/lib/button-link/button-link.component.ts
@@ -5,7 +5,7 @@ import {
   ViewEncapsulation,
 } from '@angular/core';
 
-import { ButtonCoreInputs } from '../button-core-inputs/button-core-inputs';
+import { ButtonCore } from '../button-core/button-core';
 
 @Component({
   selector: 'spy-button-link',
@@ -14,8 +14,8 @@ import { ButtonCoreInputs } from '../button-core-inputs/button-core-inputs';
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
 })
-export class ButtonLinkComponent extends ButtonCoreInputs {
-  @Input() url?: string;
-
+export class ButtonLinkComponent extends ButtonCore {
   buttonClassName = 'spy-button-link';
+
+  @Input() url?: string;
 }