+++ libs/button/src/lib/button/button.component.ts
@@ -6,7 +6,7 @@ import {
 } from '@angular/core';
 import { ToBoolean } from '@spryker/utils';
 
-import { ButtonCoreInputs } from '../button-core-inputs/button-core-inputs';
+import { ButtonCore } from '../button-core/button-core';
 
 export enum ButtonType {
   Button = 'button',
@@ -20,9 +20,10 @@ export enum ButtonType {
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class ButtonComponent extends ButtonCoreInputs {
+export class ButtonComponent extends ButtonCore {
   @Input() type: ButtonType = ButtonType.Button;
   @Input() @ToBoolean() disabled = false;
+  @Input() loading?: Boolean;
 
   buttonClassName = 'spy-button';