+++ libs/table/features/src/table-sync-state-feature/index.ts
@@ -1,2 +1,3 @@
 export * from './table-sync-state-feature.module';
 export * from './table-sync-state-feature.component';
+export * from './types';